+++ src_new/triggers/levi_ataxia/for_levi/leviticus/mnemosyne/001_Run_Start.lua
@@ -69,6 +69,7 @@ psionMindbreak = false  -- boons reset each run
 psionPsiwave = false  -- boons reset each run
 psionEarthquake = false  -- boons reset each run
 psionProphet = false  -- boons reset each run
+psionRoth = false  -- boons reset each run
 dragonMightSycaerunax = false  -- boons reset each run
 dragonRampage = false  -- boons reset each run
 dwFlashforward = false  -- boons reset each run